+++ scripts/css-macro.js
@@ -1,4 +1,5 @@
 (function () {
+    // v1.0.0
     'use strict';
 
     function mapper (args) {